Refactor the RxJS to better encapsulate each subject use case.
The subjects that now directly affect parts of the modal's functionality are now all stored and documented in the modal.ts file.
moveTo tempers with the camera position on the map
