--- client/lib/action.ts
@@ -1,5 +1,0 @@
-import { Subject } from 'rxjs';
-
-const showPopup = new Subject<boolean>();
-
-export default showPopup;
