--- client/lib/finishType.ts
@@ -1,3 +1,0 @@
-import { Subject } from 'rxjs';
-
-export default new Subject<boolean>();
