--- client/lib/inputValue.ts
@@ -1,5 +1,0 @@
-import { Subject } from 'rxjs';
-
-const input = new Subject<string>();
-
-export default input;
