--- client/lib/providedInfo.ts
@@ -1,8 +1,0 @@
-import { Subject } from 'rxjs';
-export type providedField = {
-  field: string;
-  hasProvided: boolean;
-};
-const providedInfo = new Subject<providedField>();
-
-export default providedInfo;
